fix: dedup @auth/core types, lowercase form method, add depcheck/typecheck
- Move Session/JWT module augmentation into src/types/auth.d.ts
- Add tsconfig paths mapping @auth/core to ./node_modules/@auth/core
- Fix SignOutButton form method 'POST' -> 'post' (SolidJS only accepts
  lowercase HTML form methods)
