+++ tsconfig.json
@@ -14,7 +14,9 @@
     "types": ["vinxi/types/client"],
     "isolatedModules": true,
     "paths": {
-      "~/*": ["./src/*"]
+      "~/*": ["./src/*"],
+      "@auth/core": ["./node_modules/@auth/core"],
+      "@auth/core/*": ["./node_modules/@auth/core/*"]
     }
   }
 }
